// packages/shared/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ChatType = {
  GROUP: 'GROUP',
  DIRECT: 'DIRECT'
} as const

export type ChatType = (typeof ChatType)[keyof typeof ChatType]


export const Roles = {
  admin: 'admin',
  user: 'user',
  atty: 'atty',
  statistics: 'statistics',
  notarial: 'notarial',
  archive: 'archive'
} as const

export type Roles = (typeof Roles)[keyof typeof Roles]


export const EmploymentType = {
  PERMANENT: 'PERMANENT',
  CONTRACTUAL: 'CONTRACTUAL',
  CASUAL: 'CASUAL'
} as const

export type EmploymentType = (typeof EmploymentType)[keyof typeof EmploymentType]


export const CaseType = {
  CRIMINAL: 'CRIMINAL',
  CIVIL: 'CIVIL',
  LAND_REGISTRATION_CASE: 'LAND_REGISTRATION_CASE',
  PETITION: 'PETITION',
  ELECTION: 'ELECTION',
  SCA: 'SCA',
  SHERRIFF: 'SHERRIFF',
  UNKNOWN: 'UNKNOWN'
} as const

export type CaseType = (typeof CaseType)[keyof typeof CaseType]


export const Status = {
  INACTIVE: 'INACTIVE',
  ACTIVE: 'ACTIVE',
  DEACTIVATED: 'DEACTIVATED',
  PENDING: 'PENDING',
  SUSPENDED: 'SUSPENDED',
  TERMINATED: 'TERMINATED'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const BloodType = {
  A_Positive: 'A_Positive',
  A_Negative: 'A_Negative',
  B_Positive: 'B_Positive',
  AB_Positive: 'AB_Positive',
  AB_Negative: 'AB_Negative',
  O_Positive: 'O_Positive',
  O_Negative: 'O_Negative'
} as const

export type BloodType = (typeof BloodType)[keyof typeof BloodType]


export const LogAction = {
  LOGIN_SUCCESS: 'LOGIN_SUCCESS',
  LOGIN_FAILED: 'LOGIN_FAILED',
  LOGOUT: 'LOGOUT',
  CREATE_CASE: 'CREATE_CASE',
  UPDATE_CASE: 'UPDATE_CASE',
  DELETE_CASE: 'DELETE_CASE',
  IMPORT_CASES: 'IMPORT_CASES',
  EXPORT_CASES: 'EXPORT_CASES',
  CREATE_EMPLOYEE: 'CREATE_EMPLOYEE',
  UPDATE_EMPLOYEE: 'UPDATE_EMPLOYEE',
  DELETE_EMPLOYEE: 'DELETE_EMPLOYEE',
  IMPORT_EMPLOYEES: 'IMPORT_EMPLOYEES',
  EXPORT_EMPLOYEES: 'EXPORT_EMPLOYEES',
  CREATE_USER: 'CREATE_USER',
  UPDATE_ROLE: 'UPDATE_ROLE',
  DEACTIVATE_USER: 'DEACTIVATE_USER',
  REACTIVATE_USER: 'REACTIVATE_USER',
  CHANGE_PASSWORD: 'CHANGE_PASSWORD',
  SET_INITIAL_PASSWORD: 'SET_INITIAL_PASSWORD',
  SEND_MAGIC_LINK: 'SEND_MAGIC_LINK',
  RESET_PASSWORD: 'RESET_PASSWORD',
  UPDATE_PROFILE: 'UPDATE_PROFILE',
  DELETE_USER: 'DELETE_USER'
} as const

export type LogAction = (typeof LogAction)[keyof typeof LogAction]


export const BackupFrequency = {
  MANUAL: 'MANUAL',
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
  MONTHLYe: 'MONTHLYe'
} as const

export type BackupFrequency = (typeof BackupFrequency)[keyof typeof BackupFrequency]


export const LogRetention = {
  ONE_MONTH: 'ONE_MONTH',
  THREE_MONTHS: 'THREE_MONTHS',
  SIX_MONTHS: 'SIX_MONTHS',
  ONE_YEAR: 'ONE_YEAR',
  INDEFINITELY: 'INDEFINITELY'
} as const

export type LogRetention = (typeof LogRetention)[keyof typeof LogRetention]


export const PasswordExpiration = {
  NEVER: 'NEVER',
  ONE_MONTH: 'ONE_MONTH',
  TWO_MONTHS: 'TWO_MONTHS',
  THREE_MONTHS: 'THREE_MONTHS',
  SIX_MONTHS: 'SIX_MONTHS'
} as const

export type PasswordExpiration = (typeof PasswordExpiration)[keyof typeof PasswordExpiration]


export const LockoutThreshold = {
  NONE: 'NONE',
  THREE_ATTEMPTS: 'THREE_ATTEMPTS',
  FIVE_ATTEMPTS: 'FIVE_ATTEMPTS',
  TEN_ATTEMPTS: 'TEN_ATTEMPTS'
} as const

export type LockoutThreshold = (typeof LockoutThreshold)[keyof typeof LockoutThreshold]


export const SessionTimeout = {
  FIFTEEN_MINUTES: 'FIFTEEN_MINUTES',
  THIRTY_MINUTES: 'THIRTY_MINUTES',
  ONE_HOUR: 'ONE_HOUR',
  TWO_HOURS: 'TWO_HOURS',
  FOUR_HOURS: 'FOUR_HOURS',
  EIGHT_HOURS: 'EIGHT_HOURS',
  NEVER: 'NEVER'
} as const

export type SessionTimeout = (typeof SessionTimeout)[keyof typeof SessionTimeout]